@@ -1,7 +1,6 @@
 import 'dart:async';
 import 'dart:collection';
 import 'dart:typed_data';
-import 'package:flutter/foundation.dart';
 import 'package:giphy_picker/src/model/giphy_client.dart';
 import 'package:giphy_picker/src/model/giphy_preview_types.dart';
 import 'package:giphy_picker/src/model/repository.dart';
@@ -14,24 +13,22 @@ typedef Future<GiphyCollection> GetCollection(
 /// Retrieves and caches gif collections from Giphy.
 class GiphyRepository extends Repository<GiphyGif> {
   final _client = http.Client();
-  final _previewCompleters = HashMap<int, Completer<Uint8List>>();
+  final _previewCompleters = HashMap<int, Completer<Uint8List?>>();
   final _previewQueue = Queue<int>();
   final GetCollection getCollection;
   final int maxConcurrentPreviewLoad;
-  GiphyClient _giphyClient;
+  late GiphyClient _giphyClient;
   int _previewLoad = 0;
-  final GiphyPreviewType previewType;
+  final GiphyPreviewType? previewType;
 
   GiphyRepository({
-    @required String apiKey,
-    @required this.getCollection,
+    required String apiKey,
+    required this.getCollection,
     this.maxConcurrentPreviewLoad = 4,
     int pageSize = 25,
-    ErrorListener onError,
+    ErrorListener? onError,
     this.previewType,
   }) : super(pageSize: pageSize, onError: onError) {
-    assert(getCollection != null);
-    assert(maxConcurrentPreviewLoad != null);
     _giphyClient = GiphyClient(apiKey: apiKey, client: _client);
   }
 
@@ -39,16 +36,14 @@ class GiphyRepository extends Repository<GiphyGif> {
   Future<Page<GiphyGif>> getPage(int page) async {
     final offset = page * pageSize;
     final collection = await getCollection(_giphyClient, offset, pageSize);
-    return Page(collection.data, page, collection.pagination.totalCount);
+    return Page(collection.data, page, collection.pagination?.totalCount ?? 0);
   }
 
   /// Retrieves a preview Gif image at specified index.
-  Future<Uint8List> getPreview(int index) async {
-    assert(index != null);
-
+  Future<Uint8List?> getPreview(int index) async {
     var completer = _previewCompleters[index];
     if (completer == null) {
-      completer = Completer<Uint8List>();
+      completer = Completer<Uint8List?>();
       _previewCompleters[index] = completer;
       _previewQueue.add(index);
 
@@ -61,8 +56,6 @@ class GiphyRepository extends Repository<GiphyGif> {
 
   /// Cancels retrieving specified preview image, by removing it from the queue.
   void cancelGetPreview(int index) {
-    assert(index != null);
-
     final completer = _previewCompleters.remove(index);
     if (completer != null) {
       // remove from queue
@@ -96,9 +89,9 @@ class GiphyRepository extends Repository<GiphyGif> {
     }
   }
 
-  Future<Uint8List> _loadPreviewImage(GiphyGif gif) async {
+  Future<Uint8List?> _loadPreviewImage(GiphyGif gif) async {
     // fallback to still image if preview is empty
-    String url;
+    String? url;
     switch (previewType) {
       case GiphyPreviewType.fixedWidthSmallStill:
         url = gif.images.fixedWidthSmallStill?.url;
@@ -110,7 +103,7 @@ class GiphyRepository extends Repository<GiphyGif> {
         url = gif.images.fixedHeight?.url;
         break;
       case GiphyPreviewType.original:
-        url = gif.images.original.url;
+        url = gif.images.original?.url;
         break;
       case GiphyPreviewType.previewWebp:
         url = gif.images.previewWebp?.url;
@@ -121,12 +114,15 @@ class GiphyRepository extends Repository<GiphyGif> {
       case GiphyPreviewType.originalStill:
         url = gif.images.originalStill?.url;
         break;
+      default:
+        url = null;
+        break;
     }
     if (url == null) {
-      url = gif.images.previewGif.url ??
+      url = gif.images.previewGif?.url ??
           gif.images.fixedWidthSmallStill?.url ??
           gif.images.fixedHeightDownsampled?.url ??
-          gif.images.original.url;
+          gif.images.original?.url;
     }
 
     if (url != null) {
@@ -138,11 +134,11 @@ class GiphyRepository extends Repository<GiphyGif> {
 
   /// The repository of trending gif images.
   static Future<GiphyRepository> trending({
-    @required String apiKey,
+    required String apiKey,
     String rating = GiphyRating.g,
     bool sticker = false,
-    ErrorListener onError,
-    GiphyPreviewType previewType,
+    ErrorListener? onError,
+    GiphyPreviewType? previewType,
   }) async {
     final repo = GiphyRepository(
         apiKey: apiKey,
@@ -159,13 +155,13 @@ class GiphyRepository extends Repository<GiphyGif> {
 
   /// A repository of images for given search query.
   static Future<GiphyRepository> search(
-      {@required String apiKey,
-      @required String query,
+      {required String apiKey,
+      required String query,
       String rating = GiphyRating.g,
       String lang = GiphyLanguage.english,
       bool sticker = false,
-      GiphyPreviewType previewType,
-      ErrorListener onError}) async {
+      GiphyPreviewType? previewType,
+      ErrorListener? onError}) async {
     final repo = GiphyRepository(
         apiKey: apiKey,
         previewType: previewType,
