@@ -5,11 +5,11 @@ import 'package:giphy_picker/src/widgets/giphy_image.dart';
 /// Presents a Giphy preview image.
 class GiphyPreviewPage extends StatelessWidget {
   final GiphyGif gif;
-  final Widget title;
-  final ValueChanged<GiphyGif> onSelected;
+  final Widget? title;
+  final ValueChanged<GiphyGif>? onSelected;
 
   const GiphyPreviewPage(
-      {@required this.gif, @required this.onSelected, this.title});
+      {required this.gif, required this.onSelected, this.title});
 
   @override
   Widget build(BuildContext context) {
@@ -17,7 +17,8 @@ class GiphyPreviewPage extends StatelessWidget {
 
     return Scaffold(
         appBar: AppBar(title: title, actions: <Widget>[
-          IconButton(icon: Icon(Icons.check), onPressed: () => onSelected(gif))
+          IconButton(
+              icon: Icon(Icons.check), onPressed: () => onSelected?.call(gif))
         ]),
         body: SafeArea(
             child: Center(
