@@ -5,18 +5,18 @@ import 'package:flutter/material.dart';
 /// You can also customize the Giphy Picker using `ThemeData`
 
 class GiphyDecorator {
+  /// Display or not the `AppBar`, it's [true] by default
+  final bool showAppBar;
+
+  /// Set a custom theme if you want to customize the views with your own style.
+  final ThemeData? giphyTheme;
+
+  /// Set an elevation for the search `TextField`, it's [0.0] by default
+  final double searchElevation;
+
   const GiphyDecorator({
     this.showAppBar = true,
     this.giphyTheme,
     this.searchElevation = 0.0,
   });
-
-  /// Display or not the `AppBar`, it's [true] by default
-  final bool showAppBar;
-
-  /// Set a custom theme if you want to customize the views with your own style.
-  final ThemeData giphyTheme;
-
-  /// Set an elevation for the search `TextField`, it's [0.0] by default
-  final double searchElevation;
 }
